backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ollama
pypdf
python-docx